deploy/requirements.txt
streamlit==1.32.0
langchain==0.2.0
langchain-google-genai==2.1.0
langchain-community==0.0.27
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-dotenv==1.0.1
pydantic==2.6.3
google-generativeai==0.8.3 